+++ modern-async.d.ts
@@ -220,6 +220,10 @@ declare module "timeoutPrecise" {
     export default timeoutPrecise;
     function timeoutPrecise<T>(fct: () => Promise<T> | T, amount: number): Promise<T>;
 }
+declare module "reflectStatus" {
+    export default reflectStatus;
+    function reflectStatus<T>(fct: () => Promise<T> | T): Promise<PromiseSettledResult<T>>;
+}
 declare module "modern-async" {
     export { default as asyncIterableWrap } from "asyncIterableWrap";
     export { default as asyncRoot } from "asyncRoot";
@@ -265,4 +269,5 @@ declare module "modern-async" {
     export { default as TimeoutError } from "TimeoutError";
     export { default as timeoutPrecise } from "timeoutPrecise";
     export { default as toArray } from "toArray";
+    export { default as reflectStatus } from "reflectStatus";
 }